Price update

+++ backend/db/schema.sql
@@ -6,8 +6,8 @@ CREATE TABLE IF NOT EXISTS table_types (
     slug TEXT NOT NULL UNIQUE,
     name TEXT NOT NULL,
     capacity INTEGER NOT NULL,
-    price_1h INTEGER NOT NULL,
-    price_2h INTEGER NOT NULL,
+    price_1h REAL NOT NULL,
+    price_2h REAL NOT NULL,
     sort_order INTEGER NOT NULL DEFAULT 0,
     is_active INTEGER NOT NULL DEFAULT 1,
     created_at TEXT NOT NULL DEFAULT (datetime('now'))
@@ -35,7 +35,7 @@ CREATE TABLE IF NOT EXISTS bookings (
     start_time TEXT NOT NULL,
     end_time TEXT NOT NULL,
     duration INTEGER NOT NULL CHECK (duration IN (1, 2)),
-    amount INTEGER NOT NULL,
+    amount REAL NOT NULL,
     payment_status TEXT NOT NULL DEFAULT 'Pending Verification'
         CHECK (payment_status IN ('Pending Verification', 'Payment Verified', 'Payment Rejected')),
     created_at TEXT NOT NULL DEFAULT (datetime('now')),
